Moving PlrError internal class from the Maintain___Response classes into its own separate class shared by the three of them. Resuming implementation of HDS load.
