+++ backend/app/src/main/java/com/moh/phlat/backend/esb/json/MaintainHdsResponse.java
@@ -14,6 +14,7 @@ import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
+import com.moh.phlat.backend.esb.boundary.PlrError;
 import com.moh.phlat.backend.model.ProcessData;
 
 import lombok.Getter;
@@ -53,13 +54,21 @@ public class MaintainHdsResponse implements PlrResponse {
 			
 			JsonNode root = objectMapper.readTree(json);
 			for (JsonNode ack : root.get(ACKNOWLEDGEMENTS)) {
-				if (ack.get(MSG_CODE) != null && ack.get(MSG_CODE).asText().contains(FAILED_RESPONSE_CODE)) {
-					hasError = true;
-					break;
+				if (ack.get(MSG_CODE) != null) {
+					if (ack.get(MSG_CODE).asText().contains(FAILED_RESPONSE_CODE)) {
+						hasError = true;
+					}
+					if (hasError) {
+						if (!ack.get(MSG_CODE).asText().contains(SUCCESSFUL_RESPONSE_CODE)
+								&& !ack.get(MSG_CODE).asText().contains(FAILED_RESPONSE_CODE)) {
+							logger.error("PLR returned with an error for ProcessData ID {}: {}", data.getId(), ack.get(MSG_TEXT).asText());
+							addError(ack.get(MSG_CODE).asText(), "ERROR", ack.get(MSG_TEXT).asText());
+						}
+					}
 				}
 			}
 			if (!hasError) {
-				JsonNode hds = root.get("facility");
+				JsonNode hds = root.get("provider");
 				if (hds.get("facilityIdentifiers") != null && hds.get("facilityIdentifiers").findValue("identifier") != null) {
 					hdsId = hds.get("facilityIdentifiers").findValue("identifier").asText();
 				}
@@ -91,21 +100,6 @@ public class MaintainHdsResponse implements PlrResponse {
 					"An unknown error occured while trying to load this record");
 		}
 		return pass;
-	}
-	
-	public class PlrError {
-		@Getter
-		private String errorCode;
-		@Getter
-		private String errorType;
-		@Getter
-		private String errorMessage;
-		
-		public PlrError(String errorCode, String errorType, String errorMessage) {
-			this.errorCode = errorCode;
-			this.errorType = errorType;
-			this.errorMessage = errorMessage;
-		}
 	}
 	
 	private void addError(String errorCode, String errorType, String errorMessage) {
